Create README.md
Save The World
 Updated 5 minutes ago
This is intended to be a 3D open source, synchronous, interactive, world-building, world-exploring project in which all contributors and players are welcome. Our mission is to demonstrate that each person in this world can do their part to make the world a better place. The intention of the project is to design a dynamic world full of life with everything from major detailed cities to great expansive oceans and anything in between that will respond appropriately to the stress placed upon it. Primarily, our objective is to incorporate global warming, pollution, and other existential crises within the open source platform for players to explore and address as they see fit with limitless creative innovation provided by the developers. Gamifying provides opportunity for individuals to interact on a personal level for greater accountability so that our future of the world is in safe hands moving forward.